开发

+++ src/main/java/com/jason/mt/model/LoadTimeModel.java
@@ -4,9 +4,9 @@ package com.jason.mt.model;
  * Created by jason on 2017/7/24.
  */
 public class LoadTimeModel {
-    private double val_ref_load_time;
+    private int val_ref_load_time;
 
-    private double val_load_time;
+    private int val_load_time;
 
     private String ct_poi;
 
@@ -14,19 +14,19 @@ public class LoadTimeModel {
 
     private int checkin_city_id;
 
-    public double getVal_ref_load_time() {
+    public int getVal_ref_load_time() {
         return val_ref_load_time;
     }
 
-    public void setVal_ref_load_time(double val_ref_load_time) {
+    public void setVal_ref_load_time(int val_ref_load_time) {
         this.val_ref_load_time = val_ref_load_time;
     }
 
-    public double getVal_load_time() {
+    public int getVal_load_time() {
         return val_load_time;
     }
 
-    public void setVal_load_time(double val_load_time) {
+    public void setVal_load_time(int val_load_time) {
         this.val_load_time = val_load_time;
     }
